Basic list support

+++ DuckDB.NET.Bindings/NativeMethods/NativeMethods.DataChunks.cs
@@ -26,13 +26,13 @@ public partial class NativeMethods
         public static extern unsafe ulong* DuckDBVectorGetValidity(IntPtr vector);
 
         [DllImport(DuckDbLibrary, CallingConvention = CallingConvention.Cdecl, EntryPoint = "duckdb_list_vector_get_child")]
-        public static extern long DuckDBListVectorGetChild(IntPtr vector);
+        public static extern IntPtr DuckDBListVectorGetChild(IntPtr vector);
 
         [DllImport(DuckDbLibrary, CallingConvention = CallingConvention.Cdecl, EntryPoint = "duckdb_list_vector_get_size")]
         public static extern long DuckDBListVectorGetSize(IntPtr vector);
 
         [DllImport(DuckDbLibrary, CallingConvention = CallingConvention.Cdecl, EntryPoint = "duckdb_struct_vector_get_child")]
-        public static extern long DuckDBStructVectorGetChild(IntPtr vector, long index);
+        public static extern IntPtr DuckDBStructVectorGetChild(IntPtr vector, long index);
 
         [DllImport(DuckDbLibrary, CallingConvention = CallingConvention.Cdecl, EntryPoint = "duckdb_destroy_data_chunk")]
         public static extern void DuckDBDestroyDataChunk(out IntPtr chunk);
